Monte-Carlo Methods for Simulating Luminescence Phenomena.

A collection of functions to simulate luminescence production in dosimetric materials using Monte-Carlo methods. Implemented are models for delocalised, localised and tunnelling transitions. Supported stimulation modes are TL, CW-OSL, LM-OSL, LM-IRSL, and ITL (ISO-TL).

Details

Funding

The development of RLumCarlo benefited from the support by various funding bodies:

The initial work by Johannes Friedrich, Sebastian Kreutzer and Christoph Schmidt was supported by the Deutsche Forschungsgemeinschaft (DFG, 2015–2018, SCHM 3051/4-1, "Modelling quartz luminescence signal dynamics relevant for dating and dosimetry", SCHM 3051/4-1).

Later work (2018-2019) was secured through the project "ULTIMO: Unifying Luminescence Models of quartz and feldspar DAAD: Deutscher Akademischer Austauschdienst (German Academic Exchange Service). Framework: DAAD PPP USA 2018, ID: 57387041.

The work of Sebastian Kreutzer as maintainer of the package was supported by LabEx LaScArBx (ANR - n. ANR-10-LABX-52) between 2017 and 2019.

Between 2020-2022, Sebastian Kreutzer received funding from the European Union’s Horizon 2020 research and innovation programme under the Marie Skłodowska-Curie grant agreement No 844457.

From 2023, the work as a maintainer of this package by Sebastian Kreutzer is supported by the DFG Heisenberg programme (#505822867).

References

Kreutzer, S., Friedrich, J., Pagonis, V., Laag, C., Rajovic, E., Schmidt, C., 2021. RLumCarlo: Simulating Cold Light using Monte Carlo Methods. The R Journal 13, 351–365. doi:10.32614/RJ-2021-043

Pagonis, V., Friedrich, J., Discher, M., Müller-Kirschbaum, A., Schlosser, V., Kreutzer, S., Chen, R., Schmidt, C., 2019. Excited state luminescence signals from a random distribution of defects - A new Monte Carlo simulation approach for feldspar. Journal of Luminescence 207, 266–272. doi:10.1016/j.jlumin.2018.11.024

See also

Author

Johannes Friedrich, University of Bayreuth (Germany),
Sebastian Kreutzer, Institute of Geography, Heidelberg University (Germany)
Vasilis Pagonis, McDaniel College Westminster (MD, USA),
Christoph Schmidt, University of Bayreuth (Germany),
Ena Rajovic, University of Bayreuth (Germany),
Alex Roy Duncan, University of Bayreuth (Germany),
Christian Laag, Institut de Physique du Globe de Paris, Université de Paris (France)